Enable node flags pass-through
